feat: complete hardware BOM calculation system
Complete hardware BOM calculation with component-to-hardware mapping:
- Professional nested loop architecture for dynamic hardware calculation
- Real-world hardware specifications with supplier codes and pricing
- Component count extraction using reduce method
- Type-safe dynamic object property access with bracket notation

Hardware coverage includes connectors, bolts, dowels, hinges, mounting plates, shelf pins and screws with accurate quantities per component type.

Completes Issue #4: TypeScript Foundation Implementation
