+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-// Placeholder file - will be replaced with actual utility exports
-// This file ensures Git tracks the utils folder structure
-
-export type {};
+export * from './validation';
+export * from './componentBomCalc';
+export * from './hardwareBomCalc';
